Move full_speed_ahead to quests folder
