+++ scripts/quests/full_speed_ahead.lua
